requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
transformers==4.44.2
torch==2.4.0
pypdf==4.3.1
python-multipart==0.0.20
sentencepiece==0.2.0
azure-storage-blob==12.21.0
requests==2.32.3
pymupdf